Improve microphone hardware and permission check messages
This moves them to the top of the options page, adds error icons, and only shows the permission request message if the permission was not granted.  The permission query API will only work in Chrome 43, so anyone still on Chrome 33-42 will not see the messages with this version, but given the current version is 80, that is an extremely small subset of users.